# docs/swagger.yaml
basePath: /api
definitions:
...
  web_backend_internal_app_serializer.DepartmentApplicationJSON:
    properties:
      created_at:
        type: string
      creator_login:
        type: string
      department_application_id:
        type: integer
      finish_date:
        type: string
      forming_date:
        type: string
      incomplete_items_count:
        type: integer
      moderator_login:
        type: string
      status:
        type: string
      title:
        type: string
    type: object
...
  web_backend_internal_app_serializer.StatusJSON:
    properties:
      status:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: API Support
    url: http://localhost:8080
  description: API для управления заявками отделов
  license:
    name: MIT
  title: Department Application API
  version: "1.0"
paths:
...
  /department_application/{id}:
    get:
      description: Возвращает полную информацию о заявке
      parameters:
      - description: ID заявки
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Данные заявки с устройствами
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Неверный ID
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Доступ запрещен
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Заявка не найдена
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Внутренняя ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Получить заявку по ID
      tags:
      - department_applications
  /department_application/{id}/delete-department_application:
    delete:
      description: Выполняет логическое удаление заявки
      parameters:
      - description: ID заявки
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Статус удаления
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Неверный запрос
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Заявка не найдена
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Внутренняя ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Удалить заявку
      tags:
      - department_applications
  /department_application/{id}/edit-department_application:
    put:
      consumes:
      - application/json
      description: Обновляет данные заявки
      parameters:
      - description: ID заявки
        in: path
        name: id
        required: true
        type: integer
      - description: Новые данные заявки
        in: body
        name: application
        required: true
        schema:
          $ref: '#/definitions/web_backend_internal_app_serializer.DepartmentApplicationJSON'
      produces:
      - application/json
      responses:
        "200":
          description: Обновленная заявка
          schema:
            $ref: '#/definitions/web_backend_internal_app_serializer.DepartmentApplicationJSON'
        "400":
          description: Неверные данные
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Заявка не найдена
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Внутренняя ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Изменить заявку
      tags:
      - department_applications
  /department_application/{id}/finish-department_application:
    put:
      consumes:
      - application/json
      description: Изменяет статус заявки (только для модераторов)
      parameters:
      - description: ID заявки
        in: path
        name: id
        required: true
        type: integer
      - description: Новый статус (completed/rejected)
        in: body
        name: status
        required: true
        schema:
          $ref: '#/definitions/web_backend_internal_app_serializer.StatusJSON'
      produces:
      - application/json
      responses:
        "200":
          description: Результат модерации
          schema:
            $ref: '#/definitions/web_backend_internal_app_serializer.DepartmentApplicationJSON'
        "400":
          description: Неверный запрос
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Доступ запрещен
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Заявка не найдена
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Внутренняя ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Завершить заявку
      tags:
      - department_applications
  /department_application/{id}/form-department_application:
    put:
      description: Переводит заявку в статус "formed"
      parameters:
      - description: ID заявки
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Сформированная заявка
          schema:
            $ref: '#/definitions/web_backend_internal_app_serializer.DepartmentApplicationJSON'
        "400":
          description: Неверный запрос
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Заявка не найдена
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Внутренняя ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Сформировать заявку
      tags:
      - department_applications
  /department_application/all-department_applications:
    get:
      description: Возвращает заявки с возможностью фильтрации по датам и статусу
      parameters:
      - description: Начальная дата (YYYY-MM-DD)
        in: query
        name: from-date
        type: string
      - description: Конечная дата (YYYY-MM-DD)
        in: query
        name: to-date
        type: string
      - description: Статус заявки
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Список заявок
          schema:
            items:
              $ref: '#/definitions/web_backend_internal_app_serializer.DepartmentApplicationJSON'
            type: array
        "400":
          description: Неверный формат даты
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Внутренняя ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Получить список заявок
      tags:
      - department_applications
  /department_application/department_application-cart:
    get:
      description: Возвращает информацию о текущей заявке-черновике пользователя
      produces:
      - application/json
      responses:
        "200":
          description: Данные корзины или no_draft
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Внутренняя ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Получить корзину заявки
      tags:
      - department_applications
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"